refactor(mup3d) move class to write model files to a separate module. update notebooks and tests to work with this module
